requirements.txt
Package            Version
------------------ ------------
aiohttp            3.8.6
APScheduler        3.10.4
bcrypt             4.0.1
email-validator    2.1.0.post1
fastapi            0.104.1
hypercorn          0.14.4
motor              3.3.1
packaging          23.2
pip                23.3.1
pytest             7.4.3
pytest-asyncio     0.21.1
python-decouple    3.8
python-jose        3.3.0
python-multipart   0.0.6
redis              5.0.1
requests           2.31.0
setuptools         68.2.2
websockets         12.0